Presses the ctrl key every second while active
A small multi-threaded program that press ctrl once every second.
Basically the idea was to have systems that were not on the network to
have the ability to 'disable' their screen saver, as a group policy will
revert it back to having a screen saver every 15 minutes, requiring a
log in with the original user of the system, even if they weren't
available.